Fix problem where closed windows weren't being removed from Window menu.
This occured on OS X. It might be some sort of leak of NSWindows, but
if it is, it's in SDL or OS X somewhere. Instead, when closing a
window, we remove it from the Window menu, which prevents it from
showing up to the user, and hence the game from crashing when it's
selected.