Fix crash in [NSWindow center].
On OS X 10.9, SDL crashes calling the [NSWindow center] method when
launching on the non-primary screen of a multi-screen setup.

This works around that by always launching the window centered on the
primary monitor. Although this is not ideal in a multi-monitor
environment, it's better than simply crashing.

Fixes renpy/renpy#407.